Tweak buildwith in a few ways.
Use `+` to separate components in the symlinked mozconfig. This makes it
more accurate for my terminal prompt to parse out which are active to
display (since some names contain `-`).

Allow partial configs to add to the objdir. This lets me choose which
partial configs affect the objdir or not (preventing filesystem usage
bloat).